backend/requirements.txt
flask
flask-socketio
eventlet
gunicorn
openai
langchain
langchain-openai
python-dotenv
geocoder
psutil
geopy
pyautogui
google-generativeai
langchain-google-genai
edge-tts
groq
langchain-groq
pygame
sounddevice
numpy
yt-dlp
requests
keyboard
huggingface_hub
faiss-cpu
pypdf
langchain-community
langchain-text-splitters
pyserial
pywin32; sys_platform == 'win32'
duckduckgo-search
sarvamai
nest-asyncio
transformers
torch
screen-brightness-control